XWIKI-18999: getdocuments.vm does not support iso 8601 date filters
- Update to take into account recent changes.
