+++ xwiki-platform-core/xwiki-platform-web/xwiki-platform-web-templates/src/test/java/org/xwiki/web/MacrosTest.java
@@ -22,8 +22,6 @@ package org.xwiki.web;
 import java.io.StringReader;
 import java.io.StringWriter;
 import java.util.HashMap;
-import java.util.Map;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
